% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-ris.R
\name{read_ris_csv}
\alias{read_ris_csv}
\title{Read a CSV-format bibliography}
\usage{
read_ris_csv(filename)
}
\arguments{
\item{filename}{Path to a csv file.}
}
\value{
A data frame.
}
\description{
Reads a csv of records, cleans its column names, and adds a \code{label} column
if the first column does not already look like one.
}
