% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ris-shared.R
\name{ris_index}
\alias{ris_index}
\title{Generate sequential record labels}
\usage{
ris_index(string, n, sep = "_")
}
\arguments{
\item{string}{Prefix for each label. Defaults to \code{"V"}.}

\item{n}{Number of labels to generate. If a vector is supplied, its length
is used.}

\item{sep}{Separator between prefix and number. Defaults to \code{"_"}.}
}
\value{
A character vector of \code{n} labels.
}
\description{
Builds zero-padded index labels for records that cannot be given a
meaningful name from their author and year.
}
\examples{
ris_index("ref", 3)
ris_index("ref", 12)

}
